move cublic-spline functions to react-components
