+++ packages/dashboard/src/managers/robot-trajectory-manager.ts
@@ -1,6 +1,6 @@
-import { Knot } from '../util/cublic-spline';
-import TrajectorySocketManager from './trajectory-socket-manager';
+import { Knot } from 'react-components';
 import { Authenticator } from 'rmf-auth';
+import TrajectorySocketManager from './trajectory-socket-manager';
 
 // RawVelocity received from server is in this format (x, y, theta)
 export type RawVelocity = [number, number, number];
